Enhance AABB and BVH structures for improved performance
- Added surface area and center calculation methods to AABB.
- Optimized AABB's Hit method using vector operations.
- Introduced BVHNode class with memory pool management for efficient allocation.
- Updated BVHNode's Hit method for stack-based traversal.
- Added BVHPool class to manage BVHNode allocations and reduce fragmentation.
- Refactored CUDARender to utilize BVHPool and optimized material handling with SoA.
- Improved Material class for better memory usage and performance.
- Enhanced Hittable class for polymorphic behavior across different types.
- Updated Sphere class to initialize bounding box directly and optimized Hit method.
- Adjusted project files for new source files and compilation settings.
- Improved Random utility functions for better randomness in rendering.

+++ RayTracingInOneWeekend/Interval.h
@@ -24,7 +24,7 @@ class Interval
 	{
 	}
 
-	__device__ [[nodiscard]] double size() const
+	__device__ [[nodiscard]] Float size() const
 	{
 		return max - min;
 	}
